chore: Split public and internal interface
